Renamed gui.helper package to gui.employeeview
