+++ src/main/java/gui/controller/NavbarController.java
@@ -2,7 +2,7 @@ package gui.controller;
 
 import gui.SceneSwitcher;
 import gui.billingtable.BillingTableGenerator;
-import gui.helper.EmployeeViewGenerator;
+import gui.employeeview.EmployeeViewGenerator;
 
 public class NavbarController {
     public void gotoAbout() throws Exception {
